+++ tests/tinytest/test_funs_string_lazy.R
@@ -488,6 +488,20 @@ expect_equal_lazy(
     pull(foo)
 )
 
+expect_equal_lazy(
+  mutate(test, foo = str_detect(x5, fixed("."))) |>
+    pull(foo),
+  mutate(test_df, foo = str_detect(x5, fixed("."))) |>
+    pull(foo)
+)
+
+expect_equal_lazy(
+  mutate(test, foo = grepl(".", x5, fixed = TRUE)) |>
+    pull(foo),
+  mutate(test_df, foo = grepl(".", x5, fixed = TRUE)) |>
+    pull(foo)
+)
+
 expect_warning(
   mutate(test, foo = grepl("e", x1, ignore.case = TRUE)),
   "will not be used"